Add files via upload
These are follow-up sequential draws from volunteer 3-sample draws that had at least 5 parts per billion of lead.
From the City of Chicago Department of Water Management: 
The results listed on this page are from residential homes in Chicago that have requested to have their water tested for lead. The results are anonymized to the block level, so the complete address of each home is not posted. In cases where a sample tests over 15 parts per billion (the action limit set by the EPA), the Department of Water Management will contact the homeowner to discuss the results and schedule a time to resample the water and inspect the home’s plumbing and electrical wiring to better evaluate possible causes.